Add logic to remove `is-pass` tag for the month of Farvardin
In this change, logic has been added to remove the `is-pass` tag from the date element
if the selected month is "Farvardin". This improvement ensures smoother date selection
for the month of Farvardin. If the `is-pass` tag does not exist, a message is printed
to the console to indicate that no tag was found for removal.